[#253] Refactoring. Added more methods to the ExternalButtoneListener class
