+++ app/src/main/java/com/example/scoutconcordia/ExternalButtonListener.java
@@ -7,6 +7,12 @@ import com.example.scoutconcordia.Activities.MapsActivity;
 import com.google.android.gms.maps.model.BitmapDescriptor;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.GroundOverlayOptions;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.Polygon;
+import com.google.maps.android.PolyUtil;
+
+import static com.example.scoutconcordia.Directions.displaySearchResults;
+import static com.example.scoutconcordia.Directions.getDirections;
 
 public class ExternalButtonListener extends MapsActivity {
 
@@ -260,4 +266,90 @@ public class ExternalButtonListener extends MapsActivity {
             }
         });
     }
+
+    // this is the listener for the explore inside button.
+    public static void addExploreInsideButtonListener()
+    {
+        exploreInsideButton.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View v) {
+                // we want to remove the building outline from the map so we can see the indoor floor plan
+                LatLng loc = searchMarker.getPosition();  // this is the location of the marker
+
+                // we look at the list of all polygons. If the marker is within the polygon then we want to hide the polygon from the map.
+                for (Polygon poly : polygonBuildings) {
+                    if (PolyUtil.containsLocation(loc, poly.getPoints(), true))
+                    {
+                        poly.setVisible(false);  // hide the polygon
+                        searchMarker.setVisible(false);  // hide the marker
+                        removeAllFloorOverlays();
+
+                        if (poly.getTag().equals("H Building"))
+                            setHallButtonsVisibility(true);
+                        else if (poly.getTag().equals("CC Building"))
+                            setCCButtonsVisibility(true);
+                        else if (poly.getTag().equals("VE Building"))
+                            setVEButtonsVisibility(true);
+                        else if (poly.getTag().equals("VL Building"))
+                        {
+                            LatLng ve_location = new LatLng(45.458850, -73.638660);
+                            for (Polygon poly2 : polygonBuildings)
+                            {
+                                if (PolyUtil.containsLocation(ve_location, poly2.getPoints(), true))
+                                    poly2.setVisible(false);
+                            }
+                            setVLButtonsVisibility(true);
+                        }
+                        else if (poly.getTag().equals("MB Building"))
+                            setMBButtonsVisibility(true);
+                    }
+                }
+                // we want to zoom in onto the center of the building.
+                animateCamera(loc, 19.0f);
+            };
+        });
+    }
+
+    public static void addNextStepListener()
+    {
+        nextStep.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (pathMarker != null)
+                {
+                    pathMarker.remove();
+                }
+                searchResultsIndex++; //increment the search result index
+
+                if (searchResultsIndex == searchResults.size())
+                {
+                    if (needMoreDirections)
+                    {
+                        LatLng[] dest = new LatLng[searchResults.get(searchResultsIndex - 1).length];
+                        System.arraycopy(searchResults.get(searchResultsIndex - 1), 0, dest, 0, searchResults.get(searchResultsIndex - 1).length);
+                        origin = dest[dest.length - 1];
+                        startingPoint = startingBuilding + " Building";
+                        getDirections(); // call get directions again to continue getting directions
+                    } else
+                    {
+                        // we want to reset the app back to the initial state
+                        searchResultsIndex = 0;
+                        searchPath.setVisible(false);
+                        nextStep.setVisibility(View.INVISIBLE);
+                    }
+                } else if (searchResultsIndex == 0) {
+                    resetPath();  //erase the path from outdoor directions
+                    exploreInsideButton.performClick();
+                    displaySearchResults(searchResults.get(searchResultsIndex));
+                } else if (searchResultsIndex == 100) {
+                    resetPath();
+                    nextStep.setVisibility(View.INVISIBLE);  //we have reached the end of the search
+                } else
+                {
+                    displaySearchResults(searchResults.get(searchResultsIndex));
+                }
+            }
+        });
+    }
+
+
 }
